Rename custom controls project
Took 10 minutes

Took 45 seconds
