+++ DentalCore.Wpf/DentalCore.Wpf.csproj
@@ -32,7 +32,7 @@
     <ItemGroup>
         <ProjectReference Include="..\DentalCore.Data\DentalCore.Data.csproj"/>
         <ProjectReference Include="..\DentalCore.Domain\DentalCore.Domain.csproj"/>
-        <ProjectReference Include="..\DentalCore.Wpf.Controls\DentalCore.Wpf.Controls.csproj" />
+        <ProjectReference Include="..\SimpleWpfToolkit\SimpleWpfToolkit.csproj" />
     </ItemGroup>
 
     <ItemGroup>
