+++ DentalCore.Wpf/Styles/CustomControls.xaml
@@ -1,6 +1,6 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:custom="clr-namespace:DentalCore.Wpf.Controls;assembly=DentalCore.Wpf.Controls">
+                    xmlns:custom="clr-namespace:SimpleWpfToolkit;assembly=SimpleWpfToolkit">
 
     <Style TargetType="custom:PlaceholderTextBox">
         <Setter Property="CornerRadius" Value="5" />
