+++ DentalCore.Wpf/Views/Components/TreatmentSelector.xaml
@@ -5,7 +5,7 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 
              xmlns:inners="clr-namespace:DentalCore.Wpf.ViewModels.Inners"
-             xmlns:custom="clr-namespace:DentalCore.Wpf.Controls;assembly=DentalCore.Wpf.Controls"
+             xmlns:custom="clr-namespace:SimpleWpfToolkit;assembly=SimpleWpfToolkit"
              mc:Ignorable="d"
              d:DesignHeight="300" d:DesignWidth="300">
