+++ SimpleWpfToolkit/AssemblyInfo.cs
