+++ SimpleWpfToolkit/Modal.cs
@@ -2,7 +2,7 @@
 using System.Windows.Controls;
 using System.Windows.Media;
 
-namespace DentalCore.Wpf.Controls;
+namespace SimpleWpfToolkit;
 
 public class Modal : ContentControl
 {
