+++ SimpleWpfToolkit/Themes/PlaceholderTextBox.xaml
@@ -1,16 +1,16 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:converters="clr-namespace:DentalCore.Wpf.Controls.Converters"
-                    xmlns:controls="clr-namespace:DentalCore.Wpf.Controls">
+                    xmlns:simpleWpfToolkit="clr-namespace:SimpleWpfToolkit"
+                    xmlns:converters="clr-namespace:SimpleWpfToolkit.Converters">
     
     <converters:SolidBrushToColorConverter x:Key="SolidBrushToColorConverter" />
     <BooleanToVisibilityConverter x:Key="BooleanToVisibilityConverter" />
     
-    <Style TargetType="controls:PlaceholderTextBox"
+    <Style TargetType="simpleWpfToolkit:PlaceholderTextBox"
            BasedOn="{StaticResource {x:Type TextBox}}">
         <Setter Property="Template">
             <Setter.Value>
-                <ControlTemplate TargetType="controls:PlaceholderTextBox">
+                <ControlTemplate TargetType="simpleWpfToolkit:PlaceholderTextBox">
                     <Border Background="{TemplateBinding Background}"
                             BorderThickness="{TemplateBinding BorderThickness}"
                             BorderBrush="{TemplateBinding BorderBrush}"
